+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-04-14 18:32:36
-// Commit: feat(perf): add async perf stream provider
+// Updated at: 2026-04-15 19:28:19
+// Commit: perf(perf): apply RepaintBoundary on animated perf widgets
 
 class perfComponentState {
-  final String commitRef = '374';
-  final String lastMessage = 'feat(perf): add async perf stream provider';
-  final DateTime timestamp = DateTime.parse('2026-04-14T18:32:36+05:00');
+  final String commitRef = '401';
+  final String lastMessage = 'perf(perf): apply RepaintBoundary on animated perf widgets';
+  final DateTime timestamp = DateTime.parse('2026-04-15T19:28:19+05:00');
 }